Don't call notify on object creation.
Don't call notify on object creation. This slipped in with one of the pull requests. The callback should only fire when someone has used the picker. If you want to fill in an input with initial values, do so when displaying the form, don't rely on the DRP to populate form defaults. This was a breaking change for some users.