a new setting for dropdowns mode - dropdownAdjusts
when this setting is false (default) selecting new month/year from the
dropdowns doesn't change currently selected range. this is the original
behavior and corresponds to how next/previous buttons work.

We've found however that in some cases users are more comfortable
adjusting the year (or month) and clicking 'apply' - e.g. when most of
the time ranges are from 1st of the month till end of another month.
That's the behavior when this setting is true.