Adding a less version of daterangepicker-bs3.css
This is an internal I file I use and figured it could be shared. Please note this a a rough version
which was stripped down in order to be as generic as possible.

This file relies on the existence of bootstrap (v3.2.0) 'variables.less' within the scope,
meaning 'variables.less' needs to be imported during compilation.

In case a user opts to use his own values, he can uncomment the
block in the top of the file, containing the basic vars the ruleset uses.