Allowing other date formats and checking if date is valid
This allows the user to set start and end dates as a string based either on the Americanized default locale that bootstrap-daterangepicker uses or on any of the formats that moment.js accepts. It also prevents an invalid moment.js object from being set to this.startDate and this.endDate.