Do not update range on hover.
Users find the change on hover to be confusing. Because they have already made a selection, they think the dates are changing from the hover and they do not realize clicking apply will use the selected dates. Removing these lines allows updating the dates only when selected by the user.