Add identifying class to dates from prev/next month on calendar
Currently there's no way to distinguish a disabled date from one that's disabled and just from the prev/next month via css.
You can determine not disabled with `td.available.off`.   Items that are disabled and  not current month end up with the class "off off disabled" which can't be selected for specifically.   The isCustom is unable to be used to distinguish this either since it's not passed if it's the left or right calendar.