fix rangeselection: autoApply not honoured
when clicking on one of the predefined ranges, clickApply was called causing the dialog to be closed; This is not beneficial e.g. when wanting to select a precise time after the daterange.

The dialog should therefore be only closed when the user has set the `autoApply` flag to true.