Added a calendarRendered event
Added a calendarRendered event to do some custom stuff if needed after the calendar is rendered. Differs from the showcalendar event because the show calendar event only fires once, this event is fired every time the user changes the 'page' for example by using the previous or next button. This way a developer can hook into this rendering and do some custom magic (adding markers for specific dates for example)