Fix TypeError: moment is not a function
Hey, I have been using Vitejs and Vue3. I was getting the following error while using your npm library.

"TypeError: moment is not a function"

It appears when inserted into the vitejs build, the library tries to use 'moment' however it receives an object instead of moment, the object contains a 'default' property that is the actual moment method.

I've added a check when the package is built via vitejs to check if moment is not a function and whether it has a default property, and if so, override moment with moment.default and continue as usual.

This change fixed the error on our project.

Thanks in advance for this awesome library.